main/requirements.txt
matplotlib==3.9.2
numpy==2.1.3
openai==1.54.4
pandas==2.2.3
protobuf==4.25.3
python-dotenv==1.0.1
streamlit==1.38.0
streamlit_extras==0.5.0
